fixed most type conversion (#179)
* fixed most type conversion

* mismatched mission control parser

+++ subsystem-arm/implementations/routers/joint-router.hpp
@@ -24,11 +24,11 @@ namespace sjsu::arm
 
         arm_arguments SetArmArguments(arm_arguments arguments)
         {
-            rotunda_.SetAngle(units::angle::degree_t(arguments.joint_args.rotunda_angle), units::angular_velocity::revolutions_per_minute_t(arguments.joint_args.speed));
-            shoulder_.SetAngle(units::angle::degree_t(arguments.joint_args.shoulder_angle), units::angular_velocity::revolutions_per_minute_t(arguments.joint_args.speed));
-            elbow_.SetAngle(units::angle::degree_t(arguments.joint_args.elbow_angle), units::angular_velocity::revolutions_per_minute_t(arguments.joint_args.speed));
-            left_wrist_.SetAngle(units::angle::degree_t(arguments.joint_args.wrist_pitch_angle + arguments.joint_args.wrist_roll_angle), units::angular_velocity::revolutions_per_minute_t(arguments.joint_args.speed));
-            right_wrist_.SetAngle(units::angle::degree_t(arguments.joint_args.wrist_roll_angle - arguments.joint_args.wrist_pitch_angle), units::angular_velocity::revolutions_per_minute_t(arguments.joint_args.speed));
+            rotunda_.SetAngle(units::angle::degree_t(static_cast<float>(arguments.joint_args.rotunda_angle)), units::angular_velocity::revolutions_per_minute_t(static_cast<float>(arguments.joint_args.speed)));
+            shoulder_.SetAngle(units::angle::degree_t(static_cast<float>(arguments.joint_args.shoulder_angle)), units::angular_velocity::revolutions_per_minute_t(static_cast<float>(arguments.joint_args.speed)));
+            elbow_.SetAngle(units::angle::degree_t(static_cast<float>(arguments.joint_args.elbow_angle)), units::angular_velocity::revolutions_per_minute_t(static_cast<float>(arguments.joint_args.speed)));
+            left_wrist_.SetAngle(units::angle::degree_t(static_cast<float>(arguments.joint_args.wrist_pitch_angle) + static_cast<float>(arguments.joint_args.wrist_roll_angle)), units::angular_velocity::revolutions_per_minute_t(static_cast<float>(arguments.joint_args.speed)));
+            right_wrist_.SetAngle(units::angle::degree_t(static_cast<float>(arguments.joint_args.wrist_roll_angle) - static_cast<float>(arguments.joint_args.wrist_pitch_angle)), units::angular_velocity::revolutions_per_minute_t(static_cast<float>(arguments.joint_args.speed)));
             return arguments;
         }
